Progressive Web Apps - Fulfill the whole checklist
This commit allows Fervor apps to meet the requirements of
app progressive web app. The only missing piece is automatic
https redirects. I think that's better handled at the nginx
or cdn level (eg cloudflare).

This commit changes how we handle static assets. It both
makes it more retrictive and allows for static assets.
Static assets are required for PWAs due to splash screens.
We've gone ahead and added sane defaults for all that for
the default app.