Improve/fix dependency referencing in the babel build step
Under certain conditions babel would not locate plugin and preset
dependencies. I decided to mimic how babel-cli and next.js handle all
this. In doing so, the references are to specific modules instead of
strings. This means we can find the dependecies within fervor's
node_modules or the library in question.